Change score list to int

+++ src/main/java/com/google/sps/servlets/Statistics.java
@@ -19,11 +19,11 @@ import java.util.List;
 
 public class Statistics {
   // Aggregated score for a list of comments  
-  // Contains sentiment scores in the range [-1, 1] with 0.2 intervals.
-  private List<Double> aggregateValues;
+  // Contains number of sentiment scores in the range [-1, 1] with 0.2 intervals.
+  private List<int> aggregateValues;
   private double averageScore;
 
-  public Statistics(List<Double> aggregateValues, double averageScore) {
+  public Statistics(List<int> aggregateValues, double averageScore) {
     this.aggregateValues = aggregateValues;
     this.averageScore = averageScore;
   }
